requirements.txt
requests==2.22.0
python_binance==0.7.9
unicorn_binance_websocket_api==1.29.0
starlette==0.14.1
binance==0.3
aiofiles==0.7.0
python-dotenv==0.19.0
uvicorn==0.14.0